pyproject.toml
[project]
name = "photoroom-python"
version = "0.1.0"
description = "Official-like Python SDK for the PhotoRoom REST API."
authors = [{ name = "Your Name", email = "you@example.com" }]
license = "MIT"
requires-python = ">=3.9"

dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "respx>=0.20.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest-cov>=4.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["photoroom", "photoroom.endpoints"]
